+++ packages/ui/src/badge/index.tsx
@@ -3,6 +3,7 @@
  *
  * - Converted styles to CSS module.
  * - Added theme color support to `info` variant.
+ * - Supported the `icon` props to make it controllable.
  */
 
 import { info, caution, error, published } from '@wordpress/icons';
@@ -36,7 +37,7 @@ export function Badge( {
 	children,
 	...props
 }: BadgeProps & React.ComponentPropsWithoutRef< 'span' > ) {
-	const icon = contextBasedIcon( intent );
+	const icon = typeof props.icon !== 'undefined' ? props.icon : contextBasedIcon( intent );
 	const hasIcon = !! icon;
 
 	return (